Use rsync and fsevents in favor of polling
Vagrant provides a shared folder type of 'rsync' that uses rsync to
copy changed files from the host to the guest operating system.  One
nice effect of this is that inotify events are triggered for filesystem
changes (this is not the case with NFS shared folders).  As a result,
Webpack can listen for fsevents instead of polling, which reduces CPU
usage and, hopefully, solves our segfault issues.