Finish the css build pipeline
Added documentation to webpack.config.js if you want details.  We're
using PostCSS for most of the processing, with a variety of plugins.
Not sure this setup is ideal... the preprocessed css syntax ends up
being application-specific and not easily readable by others.  We should
investigate SASS or something similar for some of this functionality.